refactor: move referral capture into its own component
