--- src/modules/project/pages1/projectView/state/heroAtom.ts
@@ -1,6 +1,0 @@
-// This file previously contained atoms for project-specific hero IDs.
-// This functionality has been replaced by a global, session-based
-// referringHeroIdAtom located in src/shared/state/referralAtom.ts
-// This file can likely be deleted if no other atoms reside here.
-
-export {}
